==> notebooks/4-scraping_ladypipa.ipynb <==
# Scraping de Lapypipa

Nuestro objetivo consiste en escrapear la pagina de vestidos de Ladypipa para ello usaremos tanto Selenium como Beautiful Soup.

import sys
import os
import pandas as pd
import json

sys.path.append(os.path.abspath('../'))
from src import funciones_scraping as fs

Con Selenium accedemos a la página principal de vestidos y sacamos todos los links a los que posterioirmente vamos a acceder. Estos links los almacenamos dentro de la carpeta datos para tenerlos guardados y solo tener que ejecutar una vez la función.

links_ladypipa = fs.obtencion_links_ladypipa()

with open("../datos/scraping/links_ladypipa.json", "w") as archivo_json:
    json.dump(links_ladypipa, archivo_json)

A continuación mediande BeautifulSoup iremos recabando de cada vestido la información necesaria para crear el dataframe.

with open("../datos/scraping/links_ladypipa.json", "r") as archivo_json:
    links_ladypipa = json.load(archivo_json)

dic_vestido = fs.extraccion_info_ladypipa(links_ladypipa)

df_lady_pipa = pd.DataFrame(dic_vestido)
df_lady_pipa.drop_duplicates(inplace=True)
df_lady_pipa["precio"] = df_lady_pipa["precio"].str.replace(",",".").astype(float)

Una vez hemos comprobado que todos los campos son correctos la almacenamos.

df_lady_pipa.to_csv("../datos/dataframes/df_ladypipa.csv")

df_lady_pipa.head(3)
