==> notebooks/5-limpieza.ipynb <==
# Limpieza de los DataFrames

Una vez hemos recabado todos los datos vamos a limpiarlos y dejarlos preparados para el próximo paso que será la creación e inserción en la base de datos.

Todos los dataframes son bastante parecidos, por lo general tienen los mismos campos a excepción del dataframe de natural by Lila al cual le falta la columna de `color` y en el dataframe de forever21 hay una columna de más que es la de `stock`. La idea es hacer una tabla general para juntar todas las marcas y hacer un estudio más global, por lo que habrá que añadir o eliminar las columnas anterioirmente mencionadas. Sin embargo, como en la tabla de forever21 la columna de stock aporta un gran valor, haremos un pequeño estudio particular de esta marca, creando una tabla únicamnte para sus vestidos.

import pandas as pd

Para el dataframe de Natural by Lila añadimos la columna `color`:

df_bylila = pd.read_csv("../datos/dataframes/df_bylila.csv", index_col=0)
df_bylila["color"]="sin definir"
df_bylila_con_color = df_bylila.reindex(columns=['nombre', 'marca', 'precio', 'color', 'talla', 'categoria'])
df_bylila_con_color.to_csv("../datos/dataframes/df_bylila_con_color.csv")
df_bylila_con_color.head(2)

Para el dataframe de Forever21 lo limpiamos y creamos una versión sin la columna `stock`:

df_f21 = pd.read_csv("../datos/dataframes/df_forever21.csv", index_col=0)
df_f21["categoria"] = df_f21["categoria"].fillna("sin categoria")
df_f21["categoria"] = df_f21["categoria"].replace("maxi", "largo").replace("mini", "corto")
df_forever21 = df_f21[~(df_f21["stock"]==0)]
df_forever21.to_csv("../datos/dataframes/df_forever21_final.csv")


df_forever21 = pd.read_csv("../datos/dataframes/df_forever21_final.csv", index_col=0)
df_forever21_sin_stock = df_forever21.drop(columns="stock")
df_forever21_sin_stock.to_csv("../datos/dataframes/df_forever21_sin_stock.csv")
df_forever21_sin_stock.sample(2)

Una vez tenemos los dataframe listos los concatenamos:

df_asos = pd.read_csv("../datos/dataframes/df_asos.csv", index_col=0)
df_ladypipa = pd.read_csv("../datos/dataframes/df_ladypipa.csv", index_col=0)

df_final_vestidos = pd.concat([df_asos, df_forever21_sin_stock, df_bylila_con_color, df_ladypipa])
df_final_vestidos["categoria"] = df_final_vestidos["categoria"].fillna("sin categoria")

df_final_vestidos.to_csv("../datos/dataframes/df_final_vestidos.csv")
df_final_vestidos.sample(3)

Creamos el dataframe de las marcas y categorías para que la posterior inserción en la base de datos sea más rápida y limpia.

df_marcas = pd.DataFrame(df_final_vestidos["marca"].unique())
df_marcas.to_csv("../datos/dataframes/df_marcas.csv")

df_categorias = pd.DataFrame(df_final_vestidos["categoria"].unique())
df_categorias.to_csv("../datos/dataframes/df_categorias.csv")
