==> notebooks/6-creacion_insercion_bbdd.ipynb <==
# Creación de la base de datos e inserción.

Nuestra base de datos consta de tres tablas principales: Vestidos, Marcas y Categorías. También añadiremos una cuarta tabla que será VestidosForever21 ya que al tener el stock podremos realizar otro tipo de análisis más minucioso que para el resto de tiendas. A continuación, crearemos la base de datos y las tablas:

import pandas as pd
import sys
import os

sys.path.append(os.path.abspath('../'))
from src import funciones_bbdd as fbd
from src import queries as q

nombre_bbdd = "cuerpos_reales"

# Creamos la base de datos
fbd.crear_db(nombre_bbdd,  "admin", "postgres")

# Establecemos la conexión
conexion = fbd.establecer_conn(nombre_bbdd, "admin", "postgres")

# Creamos las tablas
conexion.cursor().execute(q.querie_creacion_marcas)
conexion.cursor().execute(q.querie_creacion_categorias)
conexion.cursor().execute(q.querie_creacion_vestidos)
conexion.cursor().execute(q.querie_creacion_vestidos_forever21)

# Insertamos datos
fbd.dbeaver_commitmany(conexion, q.query_carga_marcas, "df_marcas")
fbd.dbeaver_commitmany(conexion, q.query_carga_categorias, "df_categorias")

Para insertar la tabla de vestidos necesitamos sustituir la marca y categoría por su id:

# Creamos los diccionarios para la sustitución de los valores
cursor = conexion.cursor()
cursor.execute(""" select * from marcas""")
data = cursor.fetchall()
marca_dict = {row[1]: row[0] for row in data}  

cursor = conexion.cursor()
cursor.execute(""" select * from categorias""")
data = cursor.fetchall()
categoria_dict = {row[1]: row[0] for row in data}  

# Sustituimos los valores y guardamos el dataframe
df = pd.read_csv("../datos/dataframes/df_final_vestidos.csv", index_col=0)
df["marca"] = df["marca"].map(marca_dict)
df["categoria"] = df["categoria"].map(categoria_dict)
df.to_csv("../datos/dataframes/df_con_id_vestidos.csv")

# Insertamos los datos
fbd.dbeaver_commitmany(conexion, q.querie_carga_vestidos, "df_con_id_vestidos")

Lo mismo ocurre para la tabla de vestidos de Forever 21:

# Sustituimos los valores de marca y categoria por su id
df_f21 = pd.read_csv("../datos/dataframes/df_forever21_final.csv", index_col=0)
df_f21["marca"] = df_f21["marca"].map(marca_dict)
df_f21["categoria"] = df_f21["categoria"].map(categoria_dict)
df_f21.to_csv("../datos/dataframes/df_con_id_forever21.csv")

# Insertamos los datos
fbd.dbeaver_commitmany(conexion, q.querie_carga_vestidos_f21, "df_con_id_forever21")
